+++ src/frontend/apps/impress/src/features/docs/doc-share/api/useDeleteDocAccess.ts
@@ -66,10 +66,5 @@ export const useDeleteDocAccess = (options?: UseDeleteDocAccessOptions) => {
         void options.onSuccess(data, variables, context);
       }
     },
-    onError: (error, variables, context) => {
-      if (options?.onError) {
-        void options.onError(error, variables, context);
-      }
-    },
   });
 };
