+++ src/frontend/apps/impress/src/features/docs/doc-share/components/DocShareInvitation.tsx
@@ -1,30 +1,44 @@
 import { VariantType, useToastProvider } from '@openfun/cunningham-react';
+import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
+import { css } from 'styled-components';
 
 import {
   Box,
   DropdownMenu,
   DropdownMenuOption,
+  Icon,
   IconOptions,
+  LoadMoreText,
+  Text,
 } from '@/components';
+import { QuickSearchData, QuickSearchGroup } from '@/components/quick-search';
 import { useCunninghamTheme } from '@/cunningham';
 import { Doc, Role } from '@/docs/doc-management';
 import { User } from '@/features/auth';
 
-import { useDeleteDocInvitation, useUpdateDocInvitation } from '../api';
+import {
+  useDeleteDocInvitation,
+  useDocInvitationsInfinite,
+  useUpdateDocInvitation,
+} from '../api';
 import { Invitation } from '../types';
 
 import { DocRoleDropdown } from './DocRoleDropdown';
 import { SearchUserRow } from './SearchUserRow';
 
-type Props = {
+type DocShareInvitationItemProps = {
   doc: Doc;
   invitation: Invitation;
 };
-export const DocShareInvitationItem = ({ doc, invitation }: Props) => {
+
+const DocShareInvitationItem = ({
+  doc,
+  invitation,
+}: DocShareInvitationItemProps) => {
   const { t } = useTranslation();
   const { spacingsTokens } = useCunninghamTheme();
-  const fakeUser: User = {
+  const invitedUser: User = {
     id: invitation.email,
     full_name: invitation.email,
     email: invitation.email,
@@ -79,6 +93,7 @@ export const DocShareInvitationItem = ({ doc, invitation }: Props) => {
       disabled: !canUpdate,
     },
   ];
+
   return (
     <Box
       $width="100%"
@@ -88,7 +103,7 @@ export const DocShareInvitationItem = ({ doc, invitation }: Props) => {
       <SearchUserRow
         isInvitation={true}
         alwaysShowRight={true}
-        user={fakeUser}
+        user={invitedUser}
         right={
           <Box $direction="row" $align="center" $gap={spacingsTokens['2xs']}>
             <DocRoleDropdown
@@ -107,6 +122,88 @@ export const DocShareInvitationItem = ({ doc, invitation }: Props) => {
             )}
           </Box>
         }
+      />
+    </Box>
+  );
+};
+
+type DocShareModalInviteUserRowProps = {
+  user: User;
+};
+export const DocShareModalInviteUserRow = ({
+  user,
+}: DocShareModalInviteUserRowProps) => {
+  const { t } = useTranslation();
+  return (
+    <Box
+      $width="100%"
+      data-testid={`search-user-row-${user.email}`}
+      className="--docs--doc-share-modal-invite-user-row"
+    >
+      <SearchUserRow
+        user={user}
+        right={
+          <Box
+            className="right-hover"
+            $direction="row"
+            $align="center"
+            $css={css`
+              font-family: Arial, Helvetica, sans-serif;
+              font-size: var(--c--theme--font--sizes--sm);
+              color: var(--c--theme--colors--greyscale-400);
+            `}
+          >
+            <Text $theme="primary" $variation="800">
+              {t('Add')}
+            </Text>
+            <Icon $theme="primary" $variation="800" iconName="add" />
+          </Box>
+        }
+      />
+    </Box>
+  );
+};
+
+interface QuickSearchGroupInvitationProps {
+  doc: Doc;
+}
+
+export const QuickSearchGroupInvitation = ({
+  doc,
+}: QuickSearchGroupInvitationProps) => {
+  const { t } = useTranslation();
+  const { data, hasNextPage, fetchNextPage } = useDocInvitationsInfinite({
+    docId: doc.id,
+  });
+
+  const invitationsData: QuickSearchData<Invitation> = useMemo(() => {
+    const invitations = data?.pages.flatMap((page) => page.results) || [];
+
+    return {
+      groupName: t('Pending invitations'),
+      elements: invitations,
+      endActions: hasNextPage
+        ? [
+            {
+              content: <LoadMoreText data-testid="load-more-invitations" />,
+              onSelect: () => void fetchNextPage(),
+            },
+          ]
+        : undefined,
+    };
+  }, [data?.pages, fetchNextPage, hasNextPage, t]);
+
+  if (!invitationsData.elements.length) {
+    return null;
+  }
+
+  return (
+    <Box aria-label={t('List invitation card')}>
+      <QuickSearchGroup
+        group={invitationsData}
+        renderElement={(invitation) => (
+          <DocShareInvitationItem doc={doc} invitation={invitation} />
+        )}
       />
     </Box>
   );
